Move packages and add first test
